+++ CHANGELOG.md
@@ -9,6 +9,13 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- New `PeriodIter` class as the return value from
+  `BaseModel.iter_periods()`. This returns (index, label) pairs as before but
+  also implements a `__len__()` magic method to give the number of periods to
+  loop over (the length of the iterator).
+
 ### Changed
 
 - Added blank line between equations in Python-generated model code.